refactor(dom): split helpers.ts and formFillHelpers.ts by responsibility (#216)
Two 875-line grab-bag modules split into barrel directories so each
file handles one concern:

helpers.ts (876 lines) → helpers/
- query.ts (323)      — queryDOMElements, getDOMElements, getDomContext,
                        resolveSelector, + getDocumentRootId/unpackAttributes
- screenshot.ts (477) — capturePageScreenshot, captureElementScreenshot,
                        getElementBounds, + internal scroll-to-element primitives
- index.ts (30)       — barrel re-export (keeps '@/commands/dom/helpers/index.js'
                        working for existing callers)

formFillHelpers.ts (875 lines) → formFillHelpers/
- shared.ts (81)      — formatScriptExecutionError, escape* (used by fill + pressKey)
- fill.ts (142)       — fillElement, clickElement
- pressKey.ts (224)   — pressKeyElement + dispatchKey/synthetic events + types
- scroll.ts (239)     — scrollPage + types
- stability.ts (66)   — waitForActionStability (shared by worker handlers)
- index.ts (25)       — barrel

Two callsites updated (import path '.js' → '/index.js'). No behavior
change; 620/620 tests pass.
